Codebase Update (yes, the absolute commit message returns)

+++ Include/OsintgramCXX/App/ModHandles.hpp
@@ -26,7 +26,7 @@ namespace OsintgramCXX {
 
     struct LibraryEntry {
         std::string label;
-        long long id;
+        int id;
         std::optional<std::string> author;
         std::optional<std::string> description;
         C_OnLoadEntry handler_onLoad;
